+++ packages/apostrophe/CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## UNRELEASED
 
+### Adds
+
+
 ### Fixes
+
+* Fixes an issue where the frontend was caching stale choices for `select`, `radio` and `checkboxes` fields that were saved but no longer valid (i.e removed from the schema).
+* Fixes soft-redirect module to decode the URL path before matching historic URLs, ensuring proper handling of accents, Cyrillic, and other non-ASCII characters.
+
+### Changes
 
 * Fixes an issue where the frontend was caching stale choices for `select`, `radio` and `checkboxes`
 fields that were saved but no longer valid (i.e removed from the schema).